refactor index.ts.  Reorganized core classes into a folder
