+++ src/core/LogFactory.ts
